% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCyclic.R
\name{fitCyclic}
\alias{fitCyclic}
\title{Fits the cyclic model, using a simple linear model.}
\usage{
fitCyclic(dat, yvar = NULL, xvar = NULL, grp = NULL, cov = NULL,
  P = NULL, ymin = -1, ymax = 1, step = 0.25)
}
\arguments{
\item{yvar}{dependent variable}

\item{xvar}{time variabele, usually indicates the beeps or other numerical variable. Time format is also possible.}

\item{grp}{group variabele indicator, usually indicates the days, used in "raw" plot.}

\item{cov}{vector of names containing additional variabeles (e.g. cov = c("x1", "daynr"))}

\item{P}{the periodicity of the cycle. If NULL the maximum of xvar will be taken.}

\item{ymin, ymax, step}{parameters that control axes of the plot}
}
\value{
list containing the following elements:

parameters =      amplitude and phase

fit =             object from lm()

rawDataPlot =     plot of predicted values for all observations

meansPlot =       plot with predictions averaged over subjects,

oneCyclePlot =    plot with predictions for one cycle,
}
\description{
This function fits the cyclic model, using a simple linear model (see Verboon & Leontjevas, 2018)
}
\examples{
data("pdat")
fitCyclic(dat=pdat, yvar = y = "dependentVar", xvar = "beepnr", grp = "daynr")
}
\keyword{ESM}
\keyword{cyclic}
\keyword{model}
