% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCyclicMLA.R
\name{fitCyclicMLA}
\alias{fitCyclicMLA}
\title{Fits the cyclic model on ESM data.}
\usage{
fitCyclicMLA(dat, yvar = NULL, xvar1 = NULL, xvar2 = NULL,
  id = NULL, cov = NULL, ncycle = 1, P = NULL, P2 = NULL,
  random = "intercept")
}
\arguments{
\item{dat}{data frame containing variables}

\item{yvar}{dependent variable}

\item{xvar1}{time variabele indicator of first or only cycle}

\item{xvar2}{time variabele indicator of second (longer) cycle}

\item{id}{group (subjects, clusters) identifier}

\item{cov}{additional variabels in the model}

\item{ncycle}{is the number of cyclic processes in the model (1 = 1, otherwise = 2)}

\item{P}{is the periodicity of the first cyclic process}

\item{P2}{is the periodicity of the second cyclic process}

\item{random}{keyword indicating which terms should be random effect in the model. The following keywords are available:
"intercept"= intercept only;
"first"= cyclic parameters of first cyclic process;
"second" = cyclic parameters of second cyclic process;
"cov" = covariates only, no cyclic terms; and finally
"all" = all terms}
}
\value{
list containing the following elements:

parameters =      intercept, amplitude and phase (and other terms in the model)

fit =             object from lmer()

formula =     specification of lmer model

period =      specified periodicity of cyclic process
}
\description{
This function fits the cyclic model on ESM data, using a multilevel model (see Verboon & Leontjevas, 2018)
}
\examples{
 data("smokedat")
  model <- fitCyclicMLA(dat=smokedat,  yvar="intention", xvar1="beepnr",
  xvar2="daynr", id = "subjnr", random = "all",
  ncycle = 1, cov = c("stress", "positiveAffect"))
}
\keyword{ESM}
\keyword{cyclic}
\keyword{model}
