% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdat.R
\docType{data}
\name{pdat}
\alias{pdat}
\title{Selection of variables from ESM/EMA study on smoking lapse.}
\format{A data frame with 70 rows and 5 variables:
\describe{
  \item{positiveAffect}{perceived positiveAffect}
  \item{stress}{perceived stress}
  \item{intention}{intention to quit smoking}
  \item{daynr}{day number: 1-7}
  \item{beepnr}{beep number: 1-10}
  }}
\usage{
pdat
}
\description{
A dataset with values from smokedat, aggregated over subjects
}
\keyword{datasets}
