% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCyclic.R
\name{plot.fitCyclic}
\alias{plot.fitCyclic}
\title{Plots fitCyclic object}
\usage{
\method{plot}{fitCyclic}(x, type = c("raw", "means", "oneCycle"))
}
\arguments{
\item{type}{more elements of "raw","means","oneCycle"}
}
\description{
Plots fitCyclic object
}
