% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCyclicMLA.R
\name{plot.fitCyclicMLA}
\alias{plot.fitCyclicMLA}
\title{Plots fitCyclicMLA object}
\usage{
\method{plot}{fitCyclicMLA}(x, ...)
}
\arguments{
\item{...}{ymin, ymax, step can specified to control axes of the plot}
}
\value{
plots of the aggregated data with the predicted values from the model
}
\description{
Plots fitCyclicMLA object
}
