% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdecom.R
\name{Rdecom}
\alias{Rdecom}
\title{R-squared based on ANOVA decomposition in multilevel model}
\usage{
Rdecom(dat, y, grpid, cov, xvar1, xvar2 = NULL, fit)
}
\arguments{
\item{dat}{data set}

\item{y}{dependent variable of the multilevel model}

\item{grpid}{group identity (cluster variable)}

\item{cov}{covariates in the MLA model}

\item{xvar1}{name of first time variable from the cyclic process}

\item{xvar2}{name of optional second time variable from the cyclic process}

\item{fit}{lmer or lmerTest object (fit multilevel model)}
}
\value{
vector containing three values, named: "level_1_R2","level_2_R2", "total_R2"
}
\description{
Function that computes the R-squared measure based on  ANOVA decomposition measure (Edoardo Costantini, 2018).
The explained variance is split into the explained variance for level-1 (within subjects), for the level-2 (between subjects)
and the total explained variance.
}
\references{
Constantini, E. (2018). R-squared measures in Multilevel Modeling:
The undesirable property of negative R-squared values.
Downloaded from \emph{http://arno.uvt.nl/show.cgi?fid=146739}, August 21, 2019.
}
