% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCyclic.R
\name{plot.fitCyclic}
\alias{plot.fitCyclic}
\title{Plots fitCyclic object}
\usage{
\method{plot}{fitCyclic}(x, type = c("raw", "means", "combi"))
}
\arguments{
\item{x}{fitCyclic object}

\item{type}{vector indicating plot type with elements "raw","means","combi". Default is all.}
}
\description{
Plots fitCyclic object
}
