% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdecom.R
\name{Rdecom}
\alias{Rdecom}
\title{Function that computes the ANOVA decomposition measure (Edoardo Costantini, 2018)
The explained variance for level-1 (within subjects), for the level-2 (between subjects) and the total explained variance.}
\usage{
Rdecom(dat, y, grpid, predicted)
}
\arguments{
\item{dat}{data set}

\item{y}{dependent variable of the multilevel model}

\item{grpid}{group identity (cluster variable)}

\item{predicted}{predicted values from multilevel model}
}
\value{
vector containing three values, named: "level_1_R2","level_2_R2", "total_R2"
}
\description{
Function that computes the ANOVA decomposition measure (Edoardo Costantini, 2018)
The explained variance for level-1 (within subjects), for the level-2 (between subjects) and the total explained variance.
}
\references{
Constantini, E. (2018). R-squared measures in Multilevel Modelling:
The undesirable property of negative R-squared values
\emph{download from http://arno.uvt.nl/show.cgi?fid=146739}
}
